+++ OUDS/Core/Themes/Orange/Tests/Values/ComponentTokens/ThemeOverrideOfTextInputComponentTokensTests.swift
@@ -88,11 +88,6 @@ struct ThemeOverrideOfTextInputComponentTokensTests {
         #expect(inheritedTheme.textInput.textInputBorderRadiusRounded == MockThemeTextInputComponentTokenProvider.mockThemeTextInputBorderRadius)
     }
 
-    @Test func inheritedThemeCanOverrideTextInputBorderRadiusDefault() throws {
-        #expect(inheritedTheme.textInput.textInputBorderWidthDefault != abstractTheme.textInput.textInputBorderWidthDefault)
-        #expect(inheritedTheme.textInput.textInputBorderWidthDefault == MockThemeTextInputComponentTokenProvider.mockThemeTextInputBorderWidth)
-    }
-
     @Test func inheritedThemeCanOverrideTextInputBorderRadiusWidthFocus() throws {
         #expect(inheritedTheme.textInput.textInputBorderWidthFocus != abstractTheme.textInput.textInputBorderWidthFocus)
         #expect(inheritedTheme.textInput.textInputBorderWidthFocus == MockThemeTextInputComponentTokenProvider.mockThemeTextInputBorderWidth)