+++ OUDS/Core/Themes/Wireframe/Sources/Providers/SemanticTokens/WireframeThemeFontSemanticTokensProvider.swift
@@ -14,8 +14,6 @@
 import OUDS
 import OUDSFoundations
 
-// swiftlint:disable type_name
-
 /// A class which wraps all **font semantic tokens**, *multiple*, *composite* or not, and expose them.
 /// This provider should be integrated as a `AllFontSemanticTokensProvider` implementation inside `OUDSTheme` so as to provide
 /// all tokens to the users.
@@ -50,5 +48,3 @@ final class WireframeThemeFontSemanticTokensProvider: AllFontSemanticTokensProvi
     // and in Core/Themes/Wireframe/Values/SemanticTokens/WireframeTheme+FontMultipleSemanticTokens.swift,
     // This declaration of OUDSFontSemanticTokensProvider is here to allow to write documentation.
 }
-
-// swiftlint:enable type_name