+++ OUDS/Core/Themes/Wireframe/Sources/Providers/SemanticTokens/WireframeThemeSizeSemanticTokensProvider.swift
@@ -14,8 +14,6 @@
 import OUDS
 import OUDSFoundations
 
-// swiftlint:disable type_name
-
 /// A class which wraps all **size semantic tokens**, *multiple* or not, and expose them.
 /// This provider should be integrated as a `AllSizeSemanticTokensProvider` implementation inside `OUDSTheme` so as to provide
 /// all tokens to the users.
@@ -55,5 +53,3 @@ final class WireframeThemeSizeSemanticTokensProvider: AllSizeSemanticTokensProvi
     // and in Core/Themes/Wireframe/Values/SemanticTokens/WireframeTheme+SizeMultipleSemanticTokens.swift
     // This declaration of WireframeThemeSizeSemanticTokensProvider is here to allow to write documentation.
 }
-
-// swiftlint:enable type_name