+++ OUDS/Core/Themes/Wireframe/Sources/WireframeTheme.swift
@@ -13,7 +13,6 @@
 
 import Foundation
 import OUDS
-import OUDSThemesOrange
 import SwiftUI
 
 // swiftlint:disable function_body_length
@@ -53,12 +52,9 @@ import SwiftUI
 /// - Since: 0.17.0
 public final class WireframeTheme: OUDSTheme, @unchecked Sendable {
 
-    /// Flag to avoid to register severals the fonts making some errors happen
-    private nonisolated(unsafe) static var fontsAlreadyRegistered: Bool = false
-
     // MARK: - Initializers
 
-    /// Constructor of the `Wireframe theme with its own providers of tokens.
+    /// Constructor of the `Wireframe` theme with its own providers of tokens.
     public init() {
         let borders = WireframeThemeBorderSemanticTokensProvider()
         let colors = WireframeThemeColorSemanticTokensProvider()
@@ -79,7 +75,6 @@ public final class WireframeTheme: OUDSTheme, @unchecked Sendable {
         let controlItem = WireframeThemeControlItemComponentTokensProvider(sizes: sizes, borders: borders, colors: colors, spaces: spaces)
         let divider = WireframeThemeDividerComponentTokensProvider(borders: borders)
         let link = WireframeThemeLinkComponentTokensProvider(sizes: sizes, colors: colors, spaces: spaces)
-        let listItem = WireframeThemeListItemComponentTokensProvider(sizes: sizes, colors: colors, spaces: spaces)
         let pinCodeInput = WireframeThemePinCodeInputComponentTokensProvider(spaces: spaces, dimensions: dimensions)
         let quantityInput = WireframeThemeQuantityInputComponentTokensProvider(sizes: sizes, spaces: spaces)
         let radioButton = WireframeThemeRadioButtonComponentTokensProvider(sizes: sizes, borders: borders)
@@ -108,7 +103,6 @@ public final class WireframeTheme: OUDSTheme, @unchecked Sendable {
                    chip: chip,
                    controlItem: controlItem,
                    divider: divider,
-                   listItem: listItem,
                    link: link,
                    pinCodeInput: pinCodeInput,
                    quantityInput: quantityInput,
@@ -120,21 +114,11 @@ public final class WireframeTheme: OUDSTheme, @unchecked Sendable {
                    tagInput: tagInput,
                    textArea: textArea,
                    textInput: textInput,
-                   resourcesBundle: Bundle.WireframeTheme)
-
-        registerFonts()
+                   resourcesBundle: Bundle.WireframeTheme,
+                   fontFamily: WireframeBrandFontRawTokens.fontFamilyDefault)
     }
 
     deinit {}
-
-    /// Fonts are defined in Resources/Fonts in TTF files
-    private func registerFonts() {
-        if !WireframeTheme.fontsAlreadyRegistered {
-            let fonts = Bundle.module.urls(forResourcesWithExtension: "ttf", subdirectory: nil)
-            fonts?.forEach { CTFontManagerRegisterFontsForURL($0 as CFURL, .process, nil) }
-            WireframeTheme.fontsAlreadyRegistered = true
-        }
-    }
 }
 
 // swiftlint:enable function_body_length