+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Color semantic tokens (tokens library v1.4.0) (Orange-OpenSource/ouds-ios#883)
 - Border semantic tokens (tokens library v1.4.0) (Orange-OpenSource/ouds-ios#883)
 - Text input component tokens (tokens library v1.4.0) (Orange-OpenSource/ouds-ios#883)
 - Button component tokens (tokens library v1.4.0) (Orange-OpenSource/ouds-ios#883)
@@ -47,6 +48,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+- List item component tokens
 - Input text component tokens (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
 - Select component tokens (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
 - Orange theme font family raw tokens